more events

+++ P3/Gung_Cho_Telemetry/Assets/Scripts/SistemaVidas/DanhoAlJugador.cs
@@ -11,7 +11,7 @@ public class DanhoAlJugador : MonoBehaviour
 
         if (vida != null) //si detecta al jugador
         {
-            Telemetry.Instance.TrackEvent(new ReceiveDamageEvent(TelemetryEvent.EventType.RECEIVE_DMG, (int)transform.position.x, (int)transform.position.y, ReceiveDamageEvent.DamageType.ENEMY));
+            Telemetry.Instance.TrackEvent(new ReceiveDamageEvent(TelemetryEvent.EventType.RECEIVE_DMG, transform.position.x, transform.position.y, ReceiveDamageEvent.DamageType.ENEMY));
             vida.EliminaVidaJugador(); //llamamos al metodo de quitar vida del jugador    
         }
     }
